chore: :truck: Renamed app folder to sprout, which is clearer on the purpose of it
