+++ fly.toml
@@ -17,7 +17,7 @@ primary_region = "ams"
   auto_stop_machines = true
   auto_start_machines = true
   min_machines_running = 0
-  processes = ["app"]
+  processes = ["sprout"]
 
 [[vm]]
   cpu_kind = "shared"
