+++ sprout/static/css/file-drop-zone.css
