+++ sprout/templates/file-drop-zone.html
