+++ sprout/templates/navigation-bar.html
