+++ sprout/tests/tests_models.py
@@ -1,8 +1,8 @@
 """Tests for models."""
 from django.test import TestCase
 
-from app.models import ColumnMetadata, TableMetadata
-from app.tests.db_test_utils import create_metadata_table_and_column
+from sprout.models import ColumnMetadata, TableMetadata
+from sprout.tests.db_test_utils import create_metadata_table_and_column
 
 
 class MetadataTests(TestCase):
