+++ sprout/validators.py
@@ -3,7 +3,7 @@
 from django.core.validators import RegexValidator
 from django.forms import ValidationError
 
-from app.models import TableMetadata
+from sprout.models import TableMetadata
 
 
 def validate_no_special_characters(field_name: str, field_value: str) -> None:
